Render a markdown file when an action is finished
This got messy. Tried to use the new `fetch` API with the `es6-promise` and `whatwg-fetch` libraries, and just could not get typescript to be happy with it.  Had to change typescript to say we're compiling to a higher-level browser. Will need to figure that out.

Also fought and lost to react-markdown, I couldn't figure out the correct import syntax.